refactor: reduce circular imports
